ingress/app-client/evo-ingress/pyproject.toml
[tool.poetry]
name = "evo-ingress"
version = "0.1.0"
description = ""
authors = ["Moishe Keselman <team@evoeco.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.3.1"
argh = "^0.26.2"
watchdog = "^1.0.2"
pyjwt = "^1.7.0"

[tool.poetry.dev-dependencies]

[build-system]
#requires = ["poetry-core>=1.0.0"]
#build-backend = "poetry.core.masonry.api"
requires = ["setuptools >= 40.6.0", "wheel"]
build-backend = "setuptools.build_meta"


